POST Journal
Includes Validation
